requirements.txt
requests
streamlit
pandas
python-dotenv
streamlit-autorefresh